ListPreview Display Changes
The list preview completion percentages now appear on their respective list mode buttons. Additionally, the variable mode question previews now have colouring to indicate whether they are ready to be answered or not. These colourings are applied on the German and the English text area of the question so that the user may distinguish.