Fixed Input Issue
Pressing the left and the right arrow in the inputs which allow for the adding of a new question in the question preview pane would toggle between inputs. This was intended, but the reality is that it prevented using the arrow keys to navigate the string in the currently selected input; therefore, it has been disabled.